Extracted `enum ArgKind` and `enum InstrumentationArg` to `crate::arg`.
