+++ dynamic_instrumentation/src/instrument_memory.rs
@@ -20,6 +20,7 @@ use std::mem;
 use std::path::Path;
 use std::sync::Mutex;
 
+use crate::arg::{ArgKind, InstrumentationArg};
 use crate::deref::{has_outer_deref, remove_outer_deref, strip_all_deref};
 use crate::into_operand::IntoOperand;
 use crate::source::Source;
@@ -99,53 +100,6 @@ impl InstrumentMemoryOps {
         };
         let (idx, _) = self.mir_locs.lock().unwrap().insert_full(mir_loc);
         idx.try_into().unwrap()
-    }
-}
-
-#[derive(Clone, Debug)]
-enum ArgKind<'tcx> {
-    AddressUsize(Operand<'tcx>),
-    Reference(Operand<'tcx>),
-    RawPtr(Operand<'tcx>),
-}
-
-impl<'tcx> ArgKind<'tcx> {
-    pub fn inner(&self) -> &Operand<'tcx> {
-        use ArgKind::*;
-        match self {
-            AddressUsize(x) => x,
-            Reference(x) => x,
-            RawPtr(x) => x,
-        }
-    }
-
-    pub fn from_type(op: Operand<'tcx>, ty: &ty::Ty<'tcx>) -> Self {
-        use ArgKind::*;
-        (if ty.is_unsafe_ptr() {
-            RawPtr
-        } else if ty.is_region_ptr() {
-            Reference
-        } else if ty.is_integral() {
-            AddressUsize
-        } else {
-            panic!("operand is not of integer-castable type: {:?}", op)
-        })(op)
-    }
-}
-
-#[derive(Clone, Debug)]
-enum InstrumentationArg<'tcx> {
-    AddrOf(Operand<'tcx>),
-    Op(ArgKind<'tcx>),
-}
-
-impl<'tcx> InstrumentationArg<'tcx> {
-    pub fn inner(&self) -> &Operand<'tcx> {
-        use InstrumentationArg::*;
-        match self {
-            AddrOf(x) => x,
-            Op(x) => x.inner(),
-        }
     }
 }
